+++ src/tui/action.rs
@@ -61,5 +61,7 @@ make_actions! {
     ConfirmEntryRemoval "Confirm Entry Removal" Y,
     RemoveLine "Remove Line" D,
     Search "Search" SLASH,
+    OpenAllValues "Open *A*ll Values" CONTROL_A,
+    CloseAllValues "Close *A*ll Values" CONTROL_A,
 }